msm8998-common: qcom-caf: display: hwc: gpu_tonemapper: improve checkEglError()
Because eglGetError() returns only the error of the last called EGL
function (compared to glGetError() that may contain values for multiple
recorded errors), it is not necessary to do it in a loop.
Also the loop would only run for a seemingly random number of 5 times.

Also remove dead code & improve logging in checkGlError().

Test: make libgpu_tonemapper
Change-Id: I2c71bc92fa46c271b78d3171e089e2737542d770
Signed-off-by: Timi Rautamäki <timi.rautamaki@gmail.com>
Signed-off-by: Vitorgl2003 <vitorgl.2003@gmail.com>
Signed-off-by: ImPrashantt <prashant33968@gmail.com>